refactor(attestation-service): give the joiner's boot path its own module
The service plays both sides of the root-key bootstrap: it answers the
handshake for the nodes joining after it, and at boot it runs the
requester side against its peers until one hands it the key. Both lived
in server.rs, where the peer-fetch loop sat under an RPC server it has
nothing to do with.

Move the joiner's half — ensure_root_key_present, the custodian-readiness
wait, and the peer-fetch loop — into join.rs. server.rs keeps the service
struct, the two RPC impls, and start_server, which now calls one function
and moves on. Each module's doc names the other, since one process being
both sides of one handshake is what a reader needs up front.

ATTESTATION_TYPE moves to lib.rs: both roles hold to it — the evidence
this node mints for its own join, and the evidence it accepts from the
nodes joining after it — so it belongs to neither module alone.

A pure move otherwise: the relocated bodies are unchanged but for
ensure_root_key_present becoming pub(crate).
